[ANI-71-funnel] refactor: Funnel 관련 로직 하나의 훅 함수로 취합
