--- apps/web/src/shared/ui/index.ts
@@ -1,1 +1,0 @@
-export { useFunnel } from './funnel'
